backend/pyproject.toml
[project]
name = "word-workspace-backend"
version = "0.1.0"
description = "FastAPI backend for a DOCX workspace built around ONLYOFFICE."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.8",
  "httpx>=0.28.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.7.1",
  "psycopg[binary]>=3.2.5",
  "pyjwt>=2.10.1",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.37",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["app/tests"]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
